Bug fixes (#35)
* Fixed CTA Button

This should fix the CTA button

* Update index.html

* Update index.html

* Update index.html

File path for cta.js was incorrect

* Fallback Note for CTA Button

Added a Fallback note with an email for the CTA button if it isn't working for a user

* Update cta.js

Issue with dom not loading correctly

* Update index.html

Still working on getting the cta.js to inject correctly

* Update index.html

Updated favicon to have a size per best practices

* Update styles.css

Lighthouse was upset about h1 not having a defined size or was being overwritten while inside a section, article, nav, or aside. Added section to h1, h2, and p in base typography to squash the issue

* Lightouse Accesibility Fixes

Hopefully fixing the Lighthouse h1 issue and adding improved aria nav

* Update styles.css

* Update styles.css

* Update styles.css

* Update styles.css

* Update styles.css

Fixing logo on desktop

* Fix Hamburger

Cannot use JS inline anymore. Made JS file and correct script in HTML